requirements-userbot.txt
python-telegram-bot>=12.1.1
typeguard>=2.5.0
Telethon>=1.10.4
